+++ src/app/mocks/data/personal-project-response.mock.ts
@@ -1,0 +1,19 @@
+export const personalProjectsResponseMock = {
+  personalProjects: [
+    {
+      name: 'project1',
+      code: 'https://projecapi/code/project1',
+      app: 'https://projecapi/app/project1',
+    },
+    {
+      name: 'project2',
+      code: 'https://projecapi/code/project2',
+      app: 'https://projecapi/app/project2',
+    },
+    {
+      name: 'project3',
+      code: 'https://projecapi/code/project3',
+      app: 'https://projecapi/app/project3',
+    },
+  ],
+};